Fixed delete agent and delete customer queries

+++ Backend/src/main/java/com/adwProject/Backend/primary/service/agent/AgentServiceImpl.java
@@ -1,8 +1,10 @@
 package com.adwProject.Backend.primary.service.agent;
 
 import com.adwProject.Backend.primary.entity.Agent;
+import com.adwProject.Backend.primary.entity.Customer;
 import com.adwProject.Backend.primary.entity.Order;
 import com.adwProject.Backend.primary.repository.AgentRepository;
+import com.adwProject.Backend.primary.repository.CustomerRepository;
 import com.adwProject.Backend.primary.repository.OrderRepository;
 import com.adwProject.Backend.secondary.entity.User;
 import com.adwProject.Backend.secondary.repository.UserRepository;
@@ -18,6 +20,7 @@ import java.util.List;
 public class AgentServiceImpl implements AgentService {
     private final AgentRepository agentRepository;
     private final OrderRepository orderRepository;
+    private final CustomerRepository customerRepository;
     private final UserRepository userRepository;
 
     @RequestMapping(value="/primary")
@@ -36,14 +39,25 @@ public class AgentServiceImpl implements AgentService {
     @Override
     public Boolean deleteAgent(String agentCode) {
         List<Order> orders = orderRepository.findByAgentAgentCode(agentCode).orElse(null);
+        List<Customer> customers = customerRepository.findCustomerByAgentAgentCode(agentCode).orElse(null);
+
         if(orders == null) {
             throw new GraphQLException("There is no Agent according with id: " + agentCode);
         }
-        if(orders.isEmpty() && agentRepository.existsById(agentCode)) {
+
+        if(customers == null) {
+            throw new GraphQLException("There is no Customer according with id: " + agentCode);
+        }
+
+        if(orders.isEmpty() && customers.isEmpty() && agentRepository.existsById(agentCode)) {
             User user = userRepository.findById(agentCode).orElse(null);
+
             agentRepository.deleteById(agentCode);
-            user.setActive(false);
-            userRepository.save(user);
+            if(user != null) {
+                user.setActive(false);
+                userRepository.save(user);
+            }
+
             return true;
         }
         return false;
